requirements.txt
pyserial
crcmod
tabulate
fastapi
uvicorn